backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
torch>=2.3.1
timm==1.0.9
pillow>=10.4.0
numpy>=1.26.4
python-multipart==0.0.9
opencv-python==4.10.0.84

